requirements.txt
PyYAML
pre-commit
black
flake8
flake8-docstrings
isort
pep8-naming
pydantic==2.5.2
scipy==1.11.4
bayesian_optimization==1.4.3
Jetson.GPIO==2.1.5
matplotlib==3.8.2
Adafruit-SSD1306==1.6.2
Pillow==10.2.0
pyserial==3.5
luma==2.3.6
luma.oled==3.13.0
pandas==2.0.3
plotly==5.22.0
pandas==2.0.3
chart-studio==1.1.0
matplotlib==3.7.5
nbformat==5.10.4
